requirements.txt
pillow==9.2.0
tensorflow==2.9.1

